Added "How Fishing Works" Section to the Fishing Page and fixed some misinformation
misinformation that was fixed:
the Background section describes "green fish 36" as being originally worked on by GargantuanPrankster8 which is untrue, it was originally worked on by Calvin. also, the floppy fish were not added to crdg because people "asked for it", they were added because I thought it'd be funny.